nano-2.4.0

+++ src/files.c
@@ -1,4 +1,4 @@
-/* $Id: files.c 5063 2014-07-11 18:35:08Z bens $ */
+/* $Id: files.c 5118 2015-02-15 16:28:08Z bens $ */
 /**************************************************************************
  *   files.c                                                              *
  *                                                                        *
@@ -126,6 +126,7 @@ int write_lockfile(const char *lockfilename, const char *origfilename, bool modi
     pid_t mypid;
     uid_t myuid;
     struct passwd *mypwuid;
+    struct stat fileinfo;
     char *lockdata = charalloc(1024);
     char myhostname[32];
     ssize_t lockdatalen = 1024;
@@ -145,8 +146,10 @@ int write_lockfile(const char *lockfilename, const char *origfilename, bool modi
        return -1;
     }
 
-    if (delete_lockfile(lockfilename) < 0)
-        return -1;
+    /* Check if the lock exists before we try to delete it...*/
+    if (stat(lockfilename, &fileinfo) != -1)
+	if (delete_lockfile(lockfilename) < 0)
+	    return -1;
 
     if (ISSET(INSECURE_BACKUP))
         cflags = O_WRONLY | O_CREAT | O_APPEND;
@@ -156,11 +159,13 @@ int write_lockfile(const char *lockfilename, const char *origfilename, bool modi
     fd = open(lockfilename, cflags,
 	    S_IRUSR | S_IWUSR | S_IRGRP | S_IWGRP | S_IROTH | S_IWOTH);
 
-    /* Maybe we just don't have write access.  Don't stop us from
-     * opening the file at all, just don't set the lock_filename and
-     * return success. */
-    if (fd < 0 && errno == EACCES)
-        return 1;
+    /* Maybe we just don't have write access.  Print an error message
+       and continue. */
+    if (fd < 0) {
+        statusbar(_("Error writing lock file %s: %s"), lockfilename,
+		    strerror(errno));
+        return 0;
+    }
 
     /* Now we've got a safe file stream.  If the previous open() call
      * failed, this will return NULL. */
@@ -190,12 +195,12 @@ int write_lockfile(const char *lockfilename, const char *origfilename, bool modi
      * our lockfile' message in here...
      *
      * This is likely very wrong, so this is a WIP. */
-    null_at(&lockdata, lockdatalen);
+    memset(lockdata, 0, lockdatalen);
     lockdata[0] = 0x62;
     lockdata[1] = 0x30;
     lockdata[24] = mypid % 256;
     lockdata[25] = mypid / 256;
-    snprintf(&lockdata[2], 10, "nano %s", VERSION);
+    snprintf(&lockdata[2], 11, "nano %s", VERSION);
     strncpy(&lockdata[28], mypwuid->pw_name, 16);
     strncpy(&lockdata[68], myhostname, 31);
     strncpy(&lockdata[108], origfilename, 768);
@@ -229,8 +234,8 @@ int write_lockfile(const char *lockfilename, const char *origfilename, bool modi
 int delete_lockfile(const char *lockfilename)
 {
     if (unlink(lockfilename) < 0 && errno != ENOENT) {
-        statusbar(_("Error deleting lock file %s: %s"), lockfilename,
-		    strerror(errno));
+	statusbar(_("Error deleting lock file %s: %s"), lockfilename,
+		  strerror(errno));
         return -1;
     }
     return 1;
@@ -248,7 +253,7 @@ int do_lockfile(const char *filename)
 		+ strlen(locking_suffix) + 3;
     char *lockfilename = charalloc(lockfilesize);
     char *lockfiledir = NULL;
-    char lockprog[12], lockuser[16];
+    static char lockprog[11], lockuser[17];
     struct stat fileinfo;
     int lockfd, lockpid;
 
@@ -279,7 +284,7 @@ int do_lockfile(const char *filename)
             return -1;
         }
         strncpy(lockprog, &lockbuf[2], 10);
-        lockpid = lockbuf[25] * 256 + lockbuf[24];
+        lockpid = (unsigned char)lockbuf[25] * 256 + (unsigned char)lockbuf[24];
         strncpy(lockuser, &lockbuf[28], 16);
 #ifdef DEBUG
         fprintf(stderr, "lockpid = %d\n", lockpid);
@@ -288,8 +293,9 @@ int do_lockfile(const char *filename)
         fprintf(stderr, "user which created this lock file should be %s\n",
                 lockuser);
 #endif
-        sprintf(promptstr, "File being edited (by %s, PID %d, user %s), continue?",
-                              lockprog, lockpid, lockuser);
+	/* TRANSLATORS: The second %s is the name of the user, the third that of the editor. */
+	sprintf(promptstr, _("File %s is being edited (by %s with %s, PID %d); continue?"),
+			   filename, lockuser, lockprog, lockpid);
         ans = do_yesno_prompt(FALSE, promptstr);
         if (ans < 1) {
             blank_statusbar();
@@ -314,6 +320,7 @@ int do_lockfile(const char *filename)
  * necessary, and then open and read the file, if applicable. */
 void open_buffer(const char *filename, bool undoable)
 {
+    bool quiet = FALSE;
     bool new_buffer = (openfile == NULL
 #ifndef DISABLE_MULTIBUFFER
 	 || ISSET(MULTIBUFFER)
@@ -337,13 +344,31 @@ void open_buffer(const char *filename, bool undoable)
 
     /* If we're loading into a new buffer, add a new entry to
      * openfile. */
-    if (new_buffer)
+    if (new_buffer) {
 	make_new_buffer();
+
+#ifndef NANO_TINY
+	if (ISSET(LOCKING) && filename[0] != '\0') {
+	    int lockstatus = do_lockfile(filename);
+	    if (lockstatus < 0) {
+#ifndef DISABLE_MULTIBUFFER
+		if (openfile->next) {
+		    close_buffer(TRUE);
+		    return;
+		}
+#endif
+	    } else if (lockstatus == 0) {
+		quiet = TRUE;
+	    }
+ 	}
+#endif
+    }
+
 
     /* If the filename isn't blank, and we are not in NOREAD_MODE,
      * open the file.  Otherwise, treat it as a new file. */
     rc = (filename[0] != '\0' && !ISSET(NOREAD_MODE)) ?
-		open_file(filename, new_buffer, &f) : -2;
+		open_file(filename, new_buffer, quiet, &f) : -2;
 
     /* If we have a file, and we're loading into a new buffer, update
      * the filename. */
@@ -395,7 +420,7 @@ void replace_buffer(const char *filename)
 
     /* If the filename isn't blank, open the file.  Otherwise, treat it
      * as a new file. */
-    rc = (filename[0] != '\0') ? open_file(filename, TRUE, &f) : -2;
+    rc = (filename[0] != '\0') ? open_file(filename, TRUE, FALSE, &f) : -2;
 
     /* Reinitialize the text of the current buffer. */
     free_filestruct(openfile->fileage);
@@ -431,14 +456,15 @@ void display_buffer(void)
 #ifndef DISABLE_MULTIBUFFER
 /* Switch to the next file buffer if next_buf is TRUE.  Otherwise,
  * switch to the previous file buffer. */
-void switch_to_prevnext_buffer(bool next_buf)
+void switch_to_prevnext_buffer(bool next_buf, bool quiet)
 {
     assert(openfile != NULL);
 
     /* If only one file buffer is open, indicate it on the statusbar and
      * get out. */
     if (openfile == openfile->next) {
-	statusbar(_("No more open file buffers"));
+	if (quiet == FALSE)
+	    statusbar(_("No more open file buffers"));
 	return;
     }
 
@@ -454,9 +480,10 @@ void switch_to_prevnext_buffer(bool next_buf)
     display_buffer();
 
     /* Indicate the switch on the statusbar. */
-    statusbar(_("Switched to %s"),
-	((openfile->filename[0] == '\0') ? _("New Buffer") :
-	openfile->filename));
+    if (quiet == FALSE)
+	statusbar(_("Switched to %s"),
+	    ((openfile->filename[0] == '\0') ? _("New Buffer") :
+	      openfile->filename));
 
 #ifdef DEBUG
     dump_filestruct(openfile->current);
@@ -467,19 +494,21 @@ void switch_to_prevnext_buffer(bool next_buf)
 /* Switch to the previous entry in the openfile filebuffer. */
 void switch_to_prev_buffer_void(void)
 {
-    switch_to_prevnext_buffer(FALSE);
+    switch_to_prevnext_buffer(FALSE, FALSE);
 }
 
 /* Switch to the next entry in the openfile filebuffer. */
 void switch_to_next_buffer_void(void)
 {
-    switch_to_prevnext_buffer(TRUE);
+    switch_to_prevnext_buffer(TRUE, FALSE);
 }
 
 /* Delete an entry from the openfile filebuffer, and switch to the one
  * after it.  Return TRUE on success, or FALSE if there are no more open
- * file buffers. */
-bool close_buffer(void)
+ * file buffers.
+ * quiet - should we print messages switching bufers
+ */
+bool close_buffer(bool quiet)
 {
     assert(openfile != NULL);
 
@@ -492,7 +521,7 @@ bool close_buffer(void)
 #endif
 
     /* Switch to the next file buffer. */
-    switch_to_next_buffer_void();
+     switch_to_prevnext_buffer(TRUE, quiet);
 
     /* Close the file buffer we had open before. */
     unlink_opennode(openfile->prev);
@@ -888,10 +917,10 @@ void read_file(FILE *f, int fd, const char *filename, bool undoable, bool checkw
  * Return -2 if we say "New File", -1 if the file isn't opened, and the
  * fd opened otherwise.  The file might still have an error while reading
  * with a 0 return value.  *f is set to the opened file. */
-int open_file(const char *filename, bool newfie, FILE **f)
+int open_file(const char *filename, bool newfie, bool quiet, FILE **f)
 {
     struct stat fileinfo, fileinfo2;
-    int fd, quiet = 0;
+    int fd;
     char *full_filename;
 
     assert(filename != NULL && f != NULL);
@@ -905,16 +934,6 @@ int open_file(const char *filename, bool newfie, FILE **f)
 	|| (stat(full_filename, &fileinfo) == -1 && stat(filename, &fileinfo2) != -1))
 	full_filename = mallocstrcpy(NULL, filename);
 
-
-#ifndef NANO_TINY
-    if (ISSET(LOCKING)) {
-	int lockstatus = do_lockfile(full_filename);
-        if (lockstatus < 0)
-	    return -1;
-	else if (lockstatus == 0)
-	    quiet = 1;
-    }
-#endif
 
     if (stat(full_filename, &fileinfo) == -1) {
 	/* Well, maybe we can open the file even if the OS says it's
@@ -1026,7 +1045,6 @@ void do_insertfile(
     size_t current_x_save = openfile->current_x;
     ssize_t current_y_save = openfile->current_y;
     bool edittop_inside = FALSE;
-    const sc *s;
 #ifndef NANO_TINY
     bool right_side_up = FALSE, single_line = FALSE;
 #endif
@@ -1084,14 +1102,13 @@ void do_insertfile(
 	    break;
 	} else {
 	    size_t pww_save = openfile->placewewant;
+	    functionptrtype func = func_from_key(&i);
 
 	    ans = mallocstrcpy(ans, answer);
 
-	    s = get_shortcut(&i);
-
 #ifndef NANO_TINY
 #ifndef DISABLE_MULTIBUFFER
-	    if (s && s->scfunc == new_buffer_void) {
+	    if (func == new_buffer_void) {
 		/* Don't allow toggling if we're in view mode. */
 		if (!ISSET(VIEW_MODE))
 		    TOGGLE(MULTIBUFFER);
@@ -1100,14 +1117,14 @@ void do_insertfile(
 		continue;
 	    }
 #endif
-	    if (s && s->scfunc == flip_execute_void) {
+	    if (func == flip_execute_void) {
 		execute = !execute;
 		continue;
 	    }
 #endif /* !NANO_TINY */
 
 #ifndef DISABLE_BROWSER
-	    if (s && s->scfunc == to_files_void) {
+	    if (func == to_files_void) {
 		char *tmp = do_browse_from(answer);
 
 		if (tmp == NULL)
@@ -1131,18 +1148,18 @@ void do_insertfile(
 		continue;
 
 #ifndef NANO_TINY
-		/* Keep track of whether the mark begins inside the
-		 * partition and will need adjustment. */
-		if (openfile->mark_set) {
-		    filestruct *top, *bot;
-		    size_t top_x, bot_x;
+	    /* Keep track of whether the mark begins inside the
+	     * partition and will need adjustment. */
+	    if (openfile->mark_set) {
+		filestruct *top, *bot;
+		size_t top_x, bot_x;
 
-		    mark_order((const filestruct **)&top, &top_x,
+		mark_order((const filestruct **)&top, &top_x,
 			(const filestruct **)&bot, &bot_x,
 			&right_side_up);
 
-		    single_line = (top == bot);
-		}
+		single_line = (top == bot);
+	    }
 #endif
 
 #ifndef DISABLE_MULTIBUFFER
@@ -2234,7 +2251,6 @@ bool do_writeout(bool exiting)
     static bool did_credits = FALSE;
 #endif
     bool retval = FALSE;
-    const sc *s;
 
     currmenu = MWRITEFILE;
 
@@ -2306,11 +2322,12 @@ bool do_writeout(bool exiting)
 	    retval = FALSE;
 	    break;
 	} else {
+	    functionptrtype func = func_from_key(&i);
+
 	    ans = mallocstrcpy(ans, answer);
-	    s = get_shortcut(&i);
 
 #ifndef DISABLE_BROWSER
-	    if (s && s->scfunc == to_files_void) {
+	    if (func == to_files_void) {
 		char *tmp = do_browse_from(answer);
 
 		if (tmp == NULL)
@@ -2322,26 +2339,26 @@ bool do_writeout(bool exiting)
 	    } else
 #endif /* !DISABLE_BROWSER */
 #ifndef NANO_TINY
-	    if (s && s->scfunc == dos_format_void) {
+	    if (func == dos_format_void) {
 		openfile->fmt = (openfile->fmt == DOS_FILE) ? NIX_FILE :
 			DOS_FILE;
 		continue;
-	    } else if (s && s->scfunc == mac_format_void) {
+	    } else if (func == mac_format_void) {
 		openfile->fmt = (openfile->fmt == MAC_FILE) ? NIX_FILE :
 			MAC_FILE;
 		continue;
-	    } else if (s && s->scfunc == backup_file_void) {
+	    } else if (func == backup_file_void) {
 		TOGGLE(BACKUP_FILE);
 		continue;
 	    } else
 #endif /* !NANO_TINY */
-	    if (s && s->scfunc == prepend_void) {
+	    if (func == prepend_void) {
 		append = (append == PREPEND) ? OVERWRITE : PREPEND;
 		continue;
-	    } else if (s && s->scfunc == append_void) {
+	    } else if (func == append_void) {
 		append = (append == APPEND) ? OVERWRITE : APPEND;
 		continue;
-	    } else if (s && s->scfunc == do_help_void) {
+	    } else if (func == do_help_void) {
 		continue;
 	    }
